FoodLogController updated to have all functionality, LogTableViewController no longer does anything
